move onBeforeRequest handler to a separate module
